Move all files to usmcore project.
