+++ usm/usm.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "usm", "usm\usm.csproj", "{FEC9E27B-7F9D-41E3-A4C0-1571914427F9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "usmcore", "usmcore\usmcore.csproj", "{86556F7E-C22F-4A76-A16A-B4599F64CB7D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{FEC9E27B-7F9D-41E3-A4C0-1571914427F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FEC9E27B-7F9D-41E3-A4C0-1571914427F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FEC9E27B-7F9D-41E3-A4C0-1571914427F9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86556F7E-C22F-4A76-A16A-B4599F64CB7D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86556F7E-C22F-4A76-A16A-B4599F64CB7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86556F7E-C22F-4A76-A16A-B4599F64CB7D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86556F7E-C22F-4A76-A16A-B4599F64CB7D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
