+++ usm/usm/Program.cs
@@ -2,6 +2,7 @@
 using System;
 using System.IO;
 using System.IO.Compression;
+using usmcore;
 
 
 namespace usm
