+++ usm/usmcore/ReadmeCreator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 
-namespace usm
+namespace usmcore
 {
     public static class ReadmeCreator
     {
