+++ usm/usmcore/UsmJsonHelper.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace usm
+namespace usmcore
 {
 
     /// <summary>
